refactor(DST-1587): extract shared useWheelScrollX hook for Tabs and SegmentedControl
Move the wheel-to-horizontal-scroll handler out of TabList into a reusable
useWheelScrollX hook and adopt it in both Tabs (no behavior change) and
SegmentedControl (new wheel support). The scroll-trap, deltaMode-1
normalization, ctrlKey zoom bypass, and instant-behavior logic now live in
one place. Wheel unit tests move to the hook; SegmentedControl gains wheel
coverage.
